UPSTREAM: UefiPayloadPkg: Do not allocate hobs from reserved memory
With cerntain combinations of memory when special-purpose memory shows up first during the parsing, it is used for HOB creation and other UEFI memory map requirements which is not desired. While special-purpose memory is supposed to be denoted as reserved-memory which should ideally skip it in the HOB creation process, it does contain a @memory node within which unexpectedly gets construed as a regular "available" memory node. So now we check if the parent node is reserved node, do not use it for HOB.